move dataset loading to utils.py, prepare PianoComposerDataset for training
